updated branch controller authorization

+++ app/controllers/branches_controller.rb
@@ -1,11 +1,18 @@
 class BranchesController < ApplicationController
+
+  before_action = :user_logged!
+
   def show
     @branch = Branch.find(params[:id])
+    self.action_description ="show the branch with id"
+    authorize @branch
   end
 
   def create
     @agency = Agency.find(params[:agency_id])
+    self.action_description ="create a branch"
     @branch = Branch.new
+    authorize @branch
     @branch.assign_attributes(branch_params)
     @agency.branches << @branch
     if @branch.valid?
@@ -20,16 +27,22 @@ class BranchesController < ApplicationController
   def new
     @agency = Agency.this_agency(current_user)
     @branch = Branch.new
+    self.action_description ="create a branch"
+    authorize @branch
     @branch.build_address
   end
 
   def edit
     @branch = Branch.find(params[:id])
+    self.action_description ="edit the branch"
+    authorize @branch
     @branch.build_address unless @branch.address
   end
 
   def update
     @branch = Branch.find(params[:id])
+    self.action_description ="update the branch"
+    authorize @branch
     @branch.assign_attributes(branch_params)
     if @branch.valid?
       @branch.save
@@ -42,6 +55,8 @@ class BranchesController < ApplicationController
 
   def destroy
     branch = Branch.find(params[:id])
+    self.action_description ="destroy the branch"
+    authorize branch
     branch.destroy
     flash[:notice] = "Branch '#{branch.code}' deleted."
     redirect_to agency_admin_home_path
